Actually fix linting errors
