+++ src/components/grpc/GrpcContext.tsx
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+export interface GrpcContextProps {
+  baseUrl: string;
+}
+
+export const GrpcContext = createContext<GrpcContextProps | null>(null);
