+++ src/components/grpc/GrpcProvider.tsx
@@ -1,17 +1,11 @@
-import { GrpcWebFetchTransport } from "@protobuf-ts/grpcweb-transport";
-import { createContext, PropsWithChildren, useContext } from "react";
+import { PropsWithChildren } from "react";
+import { GrpcContext, GrpcContextProps } from "./GrpcContext";
 
-export interface GrpcContextProps {
-  baseUrl: string;
-}
-
-const GrpcContext = createContext<GrpcContextProps | null>(null);
-
-export interface GrpcProviderProps {
+interface GrpcProviderProps {
   context: GrpcContextProps | null;
 }
 
-function GrpcProvider({
+export default function GrpcProvider({
   children,
   context,
 }: PropsWithChildren<GrpcProviderProps>) {
@@ -19,12 +13,3 @@ function GrpcProvider({
     <GrpcContext.Provider value={context}>{children}</GrpcContext.Provider>
   );
 }
-
-function useGrpc() {
-  const context = useContext(GrpcContext);
-  return context
-    ? new GrpcWebFetchTransport({ baseUrl: context.baseUrl })
-    : null;
-}
-
-export { GrpcProvider, useGrpc };
